# public/res/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "description": "An api for interacting with an E-Commerce server.\n\n### Admin user info:\n##### Username | 'admin'\n##### Password | 'password'",
    "version": "1.0.0",
    "title": "Online E-Commerce API",
    "contact": {
      "email": "arigorn15@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "http://localhost:3000/",
...
  "schemes": ["https", "http"],
  "paths": {
...
    "/products": {
      "summary": "Product information",
      "description": "Search for products",
      "get": {
        "summary": "Search for products",
        "description": "Search the database by category_name, product_name and product_description and return results.  If search term is not provided, returns all products.  Can be filtered by price ",
        "operationId": "search_products",
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "The search term for the product or category to search for"
          },
          {
            "in": "query",
            "name": "categoryId",
            "schema": {
              "type": "integer"
            },
            "description": "The id of the category to search for"
          },
          {
            "in": "query",
            "name": "price_less_than",
            "schema": {
              "type": "number"
            },
            "description": "Search for products with a price less than this"
          },
          {
            "in": "query",
            "name": "price_greater_than",
            "schema": {
              "type": "number"
            },
            "description": "Search for products with a price greater than this"
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response",
            "content": {
              "application/json": {
                "example": [
                  {
                    "id": "00a725ab-9638-4a4b-b2da-449ec0201b07",
                    "name": "example_product1",
                    "price": "$5.00",
                    "description": "Description",
                    "qty_in_stock": 100,
                    "image_file": "image-1709469725513.jpg"
                  },
                  {
                    "id": 2,
                    "name": "example_product2",
                    "price": "$5.00",
                    "description": "Description",
                    "qty_in_stock": 100,
                    "image_file": "image-1709469725602.jpg"
                  }
                ]
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "example": {
                  "status": 404,
                  "message": "No products found"
                }
              }
            }
          }
        },
        "tags": ["Products"]
      },
      "post": {
        "summary": "Insert new products into the database",
        "description": "Allows an administrator to add new products to the database.",
        "operationId": "post_products",
        "requestBody": {
          "description": "Product Information",
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/product_update"
              },
              "encoding": {
                "image": {
                  "contentType": "image/*"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "A successful response",
            "content": {
              "application/json": {
                "example": {
                  "message": "Product Successfully Added!"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": {
                  "status": 400,
                  "message": "Please fill in all required fields!"
                }
              }
            }
          },
          "401": {
            "description": "Not logged in",
            "content": {
              "application/json": {
                "example": {
                  "status": 401,
                  "message": "Please Log In!"
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "status": 403,
                  "message": "Permission Denied!"
                }
              }
            }
          }
        },
        "tags": ["Products"]
      }
    },
    "/products/{productId}": {
      "summary": "Product information",
      "description": "View product information",
      "get": {
        "summary": "Get product information by id",
        "description": "Request the information for a specific product",
        "operationId": "get_product",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "The id of the product.",
            "required": true,
            "schema": {
              "type": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response",
            "content": {
              "application/json": {
                "example": {
                  "id": "{productId}",
                  "name": "example_product",
                  "price": "$5.00",
                  "description": "This is a description of the product",
                  "qty_in_stock": 100,
                  "image_file": "image-1709469725513.jpg"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "example": {
                  "status": 404,
                  "message": "Product with ID {productId} not found"
                }
              }
            }
          }
        },
        "tags": ["Products"]
      },
      "put": {
        "summary": "Edit product's information",
        "description": "Allows a user with the 'admin' role to edit information for a specific product.",
        "operationId": "put_product",
        "requestBody": {
          "description": "Product Information",
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/product_update"
              },
              "encoding": {
                "image": {
                  "contentType": "image/*"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "The id of the product.",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response",
            "content": {
              "application/json": {
                "example": {
                  "message": "Product Successfully Updated!"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": {
                  "status": 400,
                  "message": "Please fill in all required fields!"
                }
              }
            }
          },
          "401": {
            "description": "Not logged in",
            "content": {
              "application/json": {
                "example": {
                  "status": 401,
                  "message": "Please Log In!"
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "status": 403,
                  "message": "Permission Denied!"
                }
              }
            }
          }
        },
        "tags": ["Products"]
      }
    },
    "/products/categories": {
      "summary": "Category information",
      "description": "View or add categories",
      "get": {
        "summary": "Get list of product categories",
        "description": "Returns a list of all product categories.",
        "operationId": "get_categories",
        "responses": {
          "200": {
            "description": "A successful response",
            "content": {
              "application/json": {
                "example": [
                  {
                    "id": 1,
                    "name": "toiletries"
                  },
                  {
                    "id": 2,
                    "name": "electronics"
                  }
                ]
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "example": {
                  "status": 404,
                  "message": "No categories found!"
                }
              }
            }
          }
        },
        "tags": ["Products", "Categories"]
      },
      "post": {
        "summary": "Add a new category",
        "description": "Allows an administrator to add new categories.",
        "operationId": "post_category",
        "requestBody": {
          "description": "Product Information",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/categories_create"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "A successful response",
            "content": {
              "application/json": {
                "example": {
                  "id": 1,
                  "name": "new_category"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": {
                  "status": 400,
                  "message": "Category {categoryName} already exists!"
                }
              }
            }
          },
          "401": {
            "description": "Not logged in",
            "content": {
              "application/json": {
                "example": {
                  "status": 401,
                  "message": "Please Log In!"
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "status": 403,
                  "message": "Permission Denied!"
                }
              }
            }
          }
        },
        "tags": ["Products"]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "product_update": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "description": {
            "type": "string"
          },
          "quantity": {
            "type": "integer"
          },
          "category_ids": {
            "type": "array",
            "items": {
              "type": "integer"
            }
          },
          "image": {
            "type": "string",
            "format": "binary"
          }
        },
        "xml": {
          "name": "product_inventory"
        }
      },
...
      "categories_create": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
...
    }
  }
}